requirements.txt
pandas>=1.3.5
pymysql>=1.1.0
hail>=0.2.135
requests>=2.25.0
tqdm>=4.60.0
lxml>=4.6.0
python-dotenv>=0.19.0
llm_utils @ git+https://github.com/bw2/llm_utils
google-cloud-bigquery>=3.0.0
pyarrow>=10.0.0
